+++ CHANGELOG.md
@@ -27,6 +27,10 @@ Released: Upcoming
 - Attention: Replaced the following functions of the class EntityDefinition with getSubTypeField:
     - getReferenceNameField
     - getReferenceEntity
+- Attention: Replaced the following functions of the class EntityDefinition with getField:
+    - getDescription
+- Attention: Replaced the following functions of the class EntityDefinition with setField:
+    - setDescription
 - Switched to a flag-sprites.com generated css sprite for the language flags
 - Updated dependencies:
     - "silex/silex": "~2.0"